backend/requirements.txt
fastapi>=0.115
uvicorn[standard]
pydantic>=2.0
watchdog>=4.0
pyyaml>=6.0
aiofiles
